+++ warehouse/clickhouse/clickhouse.go
@@ -847,7 +847,21 @@ func (ch *HandleT) AddColumns(tableName string, columnsInfo warehouseutils.Colum
 	if len(strings.TrimSpace(cluster)) > 0 {
 		clusterClause = fmt.Sprintf(`ON CLUSTER %q`, cluster)
 	}
-	sqlStatement := addColumnsSQLStatement(ch.Namespace, tableName, clusterClause, columnsInfo)
+	format := func(_ int, columnInfo warehouseutils.ColumnInfoT) string {
+		return fmt.Sprintf(`
+		ADD COLUMN IF NOT EXISTS %q %s`,
+			columnInfo.Name,
+			getClickHouseColumnTypeForSpecificTable(tableName, columnInfo.Name, rudderDataTypesMapToClickHouse[columnInfo.Type], false),
+		)
+	}
+	sqlStatement := fmt.Sprintf(`
+		ALTER TABLE
+		%q.%q %s %s;`,
+		ch.Namespace,
+		tableName,
+		clusterClause,
+		columnsInfo.JoinColumns(format, ","),
+	)
 	pkgLogger.Infof("CH: Adding column in clickhouse for ch:%s : %v", ch.Warehouse.Destination.ID, sqlStatement)
 	_, err = ch.Db.Exec(sqlStatement)
 	return
