--- warehouse/redshift/query_builder.go
@@ -1,18 +1,0 @@
-package redshift
-
-import (
-	"fmt"
-)
-
-func addColumnSQLStatement(tableName, columnName, columnType string) string {
-	return fmt.Sprintf(`
-		ALTER TABLE
-		  %v
-		ADD
-		  COLUMN %q %s
-	`,
-		tableName,
-		columnName,
-		getRSDataType(columnType),
-	)
-}
